# docs/static/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/t/{tenantCode}/admin/groups": {
      "get": {
        "operationId": "AdminGroupController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "example": "group-1"
                          },
                          "code": {
                            "type": "string",
                            "example": "engineering"
                          },
                          "name": {
                            "type": "string",
                            "example": "Engineering"
                          },
                          "parentId": {
                            "type": "string",
                            "example": "group-parent-1",
                            "nullable": true
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2026-05-25T00:00:00.000Z"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2026-05-25T00:00:00.000Z"
                          }
                        },
                        "required": [
                          "id",
                          "code",
                          "name",
                          "parentId",
                          "createdAt",
                          "updatedAt"
                        ]
                      }
                    },
                    "total": {
                      "type": "integer",
                      "example": 42
                    },
                    "page": {
                      "type": "integer",
                      "example": 1
                    },
                    "limit": {
                      "type": "integer",
                      "example": 20
                    }
                  },
                  "required": ["items", "total", "page", "limit"]
                }
              }
            }
          },
          "401": {
            "description": "Admin session is missing or invalid"
          },
          "403": {
            "description": "Admin permission is required"
          }
        },
        "security": [
          {
            "admin_session": []
          }
        ],
        "summary": "List groups",
        "tags": ["Admin Groups"],
        "description": "Group management. Group codes are tenant-unique, hierarchy cycles are forbidden, child groups are checked before deletion, and changes are audited."
      },
      "post": {
        "operationId": "AdminGroupController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "1234567890"
                    }
                  },
                  "required": ["id"]
                }
              }
            }
          },
          "401": {
            "description": "Admin session is missing or invalid"
          },
          "403": {
            "description": "Admin permission is required"
          }
        },
        "security": [
          {
            "admin_session": []
          }
        ],
        "summary": "Create group",
        "tags": ["Admin Groups"],
        "description": "Group management. Group codes are tenant-unique, hierarchy cycles are forbidden, child groups are checked before deletion, and changes are audited."
      }
    },
    "/t/{tenantCode}/admin/groups/{id}": {
      "get": {
        "operationId": "AdminGroupController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "group-1"
                    },
                    "code": {
                      "type": "string",
                      "example": "engineering"
                    },
                    "name": {
                      "type": "string",
                      "example": "Engineering"
                    },
                    "parentId": {
                      "type": "string",
                      "example": "group-parent-1",
                      "nullable": true
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2026-05-25T00:00:00.000Z"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2026-05-25T00:00:00.000Z"
                    }
                  },
                  "required": [
                    "id",
                    "code",
                    "name",
                    "parentId",
                    "createdAt",
                    "updatedAt"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Admin session is missing or invalid"
          },
          "403": {
            "description": "Admin permission is required"
          }
        },
        "security": [
          {
            "admin_session": []
          }
        ],
        "summary": "Get group",
        "tags": ["Admin Groups"],
        "description": "Group management. Group codes are tenant-unique, hierarchy cycles are forbidden, child groups are checked before deletion, and changes are audited."
      },
      "put": {
        "operationId": "AdminGroupController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGroupDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Admin session is missing or invalid"
          },
          "403": {
            "description": "Admin permission is required"
          }
        },
        "security": [
          {
            "admin_session": []
          }
        ],
        "summary": "Update group",
        "tags": ["Admin Groups"],
        "description": "Group management. Group codes are tenant-unique, hierarchy cycles are forbidden, child groups are checked before deletion, and changes are audited."
      },
      "delete": {
        "operationId": "AdminGroupController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Admin session is missing or invalid"
          },
          "403": {
            "description": "Admin permission is required"
          }
        },
        "security": [
          {
            "admin_session": []
          }
        ],
        "summary": "Delete group",
        "tags": ["Admin Groups"],
        "description": "Group management. Group codes are tenant-unique, hierarchy cycles are forbidden, child groups are checked before deletion, and changes are audited."
      }
    },
    "/t/{tenantCode}/admin/groups/{id}/roles": {
      "get": {
        "operationId": "AdminGroupController_getRoles",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "role-1"
                      },
                      "code": {
                        "type": "string",
                        "example": "USER_ADMIN"
                      },
                      "name": {
                        "type": "string",
                        "example": "User Admin"
                      },
                      "description": {
                        "type": "string",
                        "example": "Manage users",
                        "nullable": true
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-05-25T00:00:00.000Z"
                      },
                      "updatedAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-05-25T00:00:00.000Z"
                      }
                    },
                    "required": [
                      "id",
                      "code",
                      "name",
                      "description",
                      "createdAt",
                      "updatedAt"
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "Admin session is missing or invalid"
          },
          "403": {
            "description": "Admin permission is required"
          }
        },
        "security": [
          {
            "admin_session": []
          }
        ],
        "summary": "List group roles",
        "tags": ["Admin Groups"],
        "description": "Group management. Group codes are tenant-unique, hierarchy cycles are forbidden, child groups are checked before deletion, and changes are audited."
      }
    },
    "/t/{tenantCode}/admin/groups/{id}/roles/{roleId}": {
      "post": {
        "operationId": "AdminGroupController_assignRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Admin session is missing or invalid"
          },
          "403": {
            "description": "Admin permission is required"
          }
        },
        "security": [
          {
            "admin_session": []
          }
        ],
        "summary": "Assign role to group",
        "tags": ["Admin Groups"],
        "description": "Group management. Group codes are tenant-unique, hierarchy cycles are forbidden, child groups are checked before deletion, and changes are audited."
      },
      "delete": {
        "operationId": "AdminGroupController_removeRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Admin session is missing or invalid"
          },
          "403": {
            "description": "Admin permission is required"
          }
        },
        "security": [
          {
            "admin_session": []
          }
        ],
        "summary": "Remove role from group",
        "tags": ["Admin Groups"],
        "description": "Group management. Group codes are tenant-unique, hierarchy cycles are forbidden, child groups are checked before deletion, and changes are audited."
      }
    },
...
  },
  "info": {
    "title": "Auth API",
    "description": "OIDC Authorization Server API Reference",
    "version": "0.1.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/",
      "description": "Service origin"
    },
    {
      "url": "/api",
      "description": "Vite proxy"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "admin_session"
      }
    },
    "schemas": {
...
      "CreateGroupDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateGroupDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}